added read_input_oneshot, moved tests
