+++ src/lib.rs
@@ -219,6 +219,8 @@ pub const STATUS_CONV_RDY: u8 = 0b1000_0000;
 #[cfg(test)]
 mod test {
 
+    use embedded_hal_mock::eh1::i2c::{Mock as I2cMock, Transaction as I2cTransaction};
+
     const EPS: f32 = 0.0001;
     const V_MAX: f32 = 2.048;
 
@@ -254,5 +256,78 @@ mod test {
         // one bit greater than most negative value
         let data: u16 = 0b1000_0000_0000_0001;
         assert!((single_ended_rdata_to_scaled_voltage(data as i16) - -V_MAX).abs() < EPS);
+    }
+
+    const DEFAULT_CONFIG: u8 = 0b0000_0000;
+    const DEFAULT_STATUS: u8 = 0b0000_0001;
+    const DEVICE_ADDRESS: u8 = 0b0000_0000;
+
+    fn new_ads1119(transactions: &[I2cTransaction]) -> Ads1119<I2cMock> {
+        let device_address = 0;
+        Ads1119::new(I2cMock::new(transactions), device_address)
+    }
+
+    fn destroy_ads1119(device: Ads1119<I2cMock>) {
+        device.destroy().done();
+    }
+
+    #[test]
+    fn can_read_config() {
+        let mut device = new_ads1119(&[I2cTransaction::write_read(
+            DEVICE_ADDRESS,
+            vec![CmdFlags::RREG | RegSelectFlags::CONFIG],
+            vec![DEFAULT_CONFIG],
+        )]);
+        assert_eq!(device.read_config().unwrap(), 0b0000_0000);
+        destroy_ads1119(device);
+    }
+
+    #[test]
+    fn can_write_config() {
+        let value = 0_u8;
+        let mut device = new_ads1119(&[I2cTransaction::write(
+            DEVICE_ADDRESS,
+            vec![CmdFlags::WREG | RegSelectFlags::CONFIG, value],
+        )]);
+        device.write_config(value).unwrap();
+        destroy_ads1119(device);
+    }
+
+    #[test]
+    fn can_read_status() {
+        let mut device = new_ads1119(&[I2cTransaction::write_read(
+            DEVICE_ADDRESS,
+            vec![CmdFlags::RREG | RegSelectFlags::STATUS],
+            vec![DEFAULT_STATUS],
+        )]);
+        assert_eq!(device.read_status().unwrap(), 1);
+        destroy_ads1119(device);
+    }
+
+    #[test]
+    fn can_read_input_oneshot() {
+        let input = InputSelection::AN0SingleEnded;
+        let expected_output = 16383_u16;
+        let mut device = new_ads1119(&[
+            I2cTransaction::write(DEVICE_ADDRESS, vec![CmdFlags::WREG | RegSelectFlags::CONFIG, input.bits()]),
+            I2cTransaction::write(DEVICE_ADDRESS, vec![CmdFlags::START_SYNC]),
+            I2cTransaction::write_read(
+                DEVICE_ADDRESS,
+                vec![CmdFlags::RREG | RegSelectFlags::STATUS],
+                vec![DEFAULT_STATUS],
+            ),
+            I2cTransaction::write_read(
+                DEVICE_ADDRESS,
+                vec![CmdFlags::RREG | RegSelectFlags::STATUS],
+                vec![STATUS_CONV_RDY],
+            ),
+            I2cTransaction::write_read(
+                DEVICE_ADDRESS,
+                vec![CmdFlags::RDATA],
+                vec![(expected_output >> 8) as u8,expected_output as u8],
+            ),
+        ]);
+        assert_eq!(device.read_input_oneshot(&input).unwrap(),expected_output as i16);
+        destroy_ads1119(device);
     }
 }
